requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.10
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-summernote==0.8.11.6
django-taggit==1.2.0
django-toolbelt==0.0.1
gunicorn==20.0.4
psycopg2==2.7.5
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.0.1
